+++ back-cv-java/src/main/java/com/stever/jobflow/module/cvs/service/CvService.java
@@ -1,50 +1,68 @@
-package module.cvs.service;
+package com.stever.jobflow.module.cvs.service;
 
-import core.CvSchema;
+import com.stever.jobflow.core.CvSchema;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.core.FindAndModifyOptions;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-import repository.interfaces.CVRepository;
+import com.stever.jobflow.repository.interfaces.CVRepository;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Slf4j
 @Service
 public class CvService {
+
+    @Autowired
+    private MongoTemplate mongoTemplate;
+
     private final CVRepository repo; // Spring Data Mongo repository
 
-    public CvService(CVRepository repo) { this.repo = repo; }
+    public CvService(CVRepository repo) {
+        this.repo = repo;
+    }
 
     public CvSchema create(CvSchema cv, int sub) {
-        log.info("Création d'un cv pour l'utilsateur id = {}", sub );
         cv.setMemberId(sub);
         return repo.save(cv);
     }
 
+    public CvSchema update(Map<String, String> fieldsToUpdate, String id, int sub) {
+        verifyOwnership(id, sub);
+        Query query = new Query(Criteria.where("_id").is(id));
+        Update update = new Update();
+
+        fieldsToUpdate.forEach(update::set);
+
+        return mongoTemplate.findAndModify(query, update, FindAndModifyOptions.options().returnNew(true), CvSchema.class);
+    }
+
     public List<CvSchema> findAll(int sub) {
-        log.info("Récupération de tous les CVs pour l'utilisateur id={}", sub);
         return repo.findByMemberId(sub);
     }
 
     public boolean delete(String id, int sub) {
-        log.info("Suppression du cv id={}", id);
-        verify(id, sub);
+        verifyOwnership(id, sub);
         repo.deleteById(id);
         return true;
     }
 
     public CvSchema findOne(String id, int sub) {
-        log.info("Recherche du CV id={} pour l'utilisateur id={}", id, sub);
-        return verify(id, sub);
+        return verifyOwnership(id, sub);
     }
 
-    public CvSchema verify(String id, int sub) {
-        log.debug("Vérification d'accès pour le CV id={} et utilisateur id={}", id, sub);
+    public CvSchema verifyOwnership(String id, int sub) {
         Optional<CvSchema> cv = repo.findById(id);
         return cv
-                .filter(cvFound -> cvFound.getMemberId() == sub)
+                .filter(cvFound -> Integer.valueOf(cvFound.getMemberId()).equals(sub))
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Accés non authorisé"));
     }
 }
